refactor: 역 정보 파일 읽어보는 클래스 이름 변경 및 위치 이동
